backend/requirements.txt
langgraph>=0.2.60
mcp[cli]>=1.13.0
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic-ai
fastapi>=0.115.0
uvicorn>=0.30.0
gunicorn>=23.0.0
langsmith>=0.3.0
opentelemetry-exporter-otlp>=1.30.0
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.3
PyJWT>=2.10.1
passlib[bcrypt]>=1.7.4
google-auth>=2.37.0
email-validator>=2.2.0
cryptography>=43.0.0
